NetworkModal refactor
