+++ src/components/common/Layout/components/Account/components/NetworkModal/NetworkModal.tsx
@@ -4,39 +4,21 @@ import {
   ModalContent,
   ModalHeader,
   ModalOverlay,
-  SimpleGrid,
   Text,
   useBreakpointValue,
 } from "@chakra-ui/react"
 import { useWeb3React } from "@web3-react/core"
-import { WalletConnect } from "@web3-react/walletconnect"
 import { Modal } from "components/common/Modal"
 import { Web3Connection } from "components/_app/Web3ConnectionManager"
-import { Chains, supportedChains } from "connectors"
-import useToast from "hooks/useToast"
 import { useContext } from "react"
-import NetworkButton from "./components/NetworkButton"
-import requestNetworkChange from "./utils/requestNetworkChange"
+import NetworkButtonsList from "./components/NetworkButtonsList"
 
 const NetworkModal = ({ isOpen, onClose }) => {
   const { listedChainIDs } = useContext(Web3Connection)
 
   const modalSize = useBreakpointValue({ base: "lg", md: "2xl", lg: "4xl" })
 
-  const { connector, isActive } = useWeb3React()
-  const toast = useToast()
-
-  const requestManualNetworkChange = (chain) => () =>
-    toast({
-      title: "Your wallet doesn't support switching chains automatically",
-      description: `Please switch to ${chain} from your wallet manually!`,
-      status: "error",
-    })
-
-  const listedChains =
-    listedChainIDs?.length > 0
-      ? supportedChains?.filter((chain) => listedChainIDs?.includes(Chains[chain]))
-      : supportedChains
+  const { isActive } = useWeb3React()
 
   return (
     <Modal isOpen={isOpen} onClose={onClose} size={modalSize}>
@@ -53,19 +35,7 @@ const NetworkModal = ({ isOpen, onClose }) => {
               used to know your address and sign messages so each will work equally.
             </Text>
           )}
-          <SimpleGrid columns={{ md: 2, lg: 3 }} spacing={{ base: 3, md: "18px" }}>
-            {listedChains.map((chain) => (
-              <NetworkButton
-                key={chain}
-                chain={chain}
-                requestNetworkChange={
-                  connector instanceof WalletConnect
-                    ? requestManualNetworkChange(chain)
-                    : requestNetworkChange(chain, onClose)
-                }
-              />
-            ))}
-          </SimpleGrid>
+          <NetworkButtonsList manualNetworkChangeCallback={onClose} />
         </ModalBody>
       </ModalContent>
     </Modal>
