--- src/components/common/Layout/components/Account/components/NetworkModal/components/NetworkButton/index.ts
@@ -1,3 +1,0 @@
-import NetworkButton from "./NetworkButton"
-
-export default NetworkButton
